#54: rename store
