+++ src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css'
 import { Provider } from 'react-redux'
-import { store } from './store/index'
+import { store } from './store/apis'
 import { Route, Routes } from 'react-router-dom'
 import { Container } from 'react-bootstrap'
 import Home from './routes/Home'
@@ -12,8 +12,6 @@ import MyPage from './routes/MyPage'
 import Recommended from './routes/Recommended'
 import Cart from './routes/Cart'
 import NotFound from './components/layout/NotFound'
-// import Product from './routes/Product'
-import Button from './components/common/Button'
 
 function App() {
   return (
@@ -24,7 +22,6 @@ function App() {
         <Route path="login" element={<Login />} />
         <Route path="signup" element={<SignUp />} />
         <Route path="/favorites" element={<Favorites />} />
-        {/* <Route path="/prdoucts" element={<Product />} /> */}
         <Route path="/mypage" element={<MyPage />} />
         <Route path="/recommened" element={<Recommended />} />
         <Route path="/cart" element={<Cart />} />
