+++ src/routes/Home/index.jsx
@@ -1,18 +1,10 @@
 import React from 'react'
 import { Outlet } from 'react-router-dom'
-import { useGetProductsQuery } from '../../api/productApiSlice'
+import { useGetProductsQuery } from '../../store/slices/productApiSlice'
 import Card from '../../components/common/Card'
 
 function Home() {
   const { data: products, isLoading, isError } = useGetProductsQuery()
-  // const { data, error, isLoading } = useGetProductsQuery(undefined, {
-  //   selectFromResult: ({ data, error, isLoading }) => ({
-  //     data: data?.card,
-  //     error,
-  //     isLoading,
-  //   }),
-  //   pollingInterval: 3000,
-  // })
 
   if (isLoading) {
     return <div>로딩중...</div>
@@ -25,7 +17,7 @@ function Home() {
   return (
     <div>
       {products.card.map((item) => (
-        <Card item={item} />
+        <Card item={item} key={item.product_id} />
       ))}
       <Outlet />
     </div>
